refactor: rename components dir to features
