+++ src/pages/login.tsx
@@ -1,3 +1,3 @@
-import { LoginPage } from "@Users/Auth/components";
+import { LoginPage } from "@Users/Auth/features";
 
 export default LoginPage;
